+++ src/main/java/com/dog_feliz/user_service/controller/NotificationController.java
@@ -4,21 +4,28 @@ import com.dog_feliz.user_service.controller.dto.NotificationRequestDto;
 import com.dog_feliz.user_service.controller.dto.NotificationResponseDto;
 import com.dog_feliz.user_service.entity.notification.NotificationEntity;
 import com.dog_feliz.user_service.service.NotificationService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpMethod;
+import com.dog_feliz.user_service.service.ValidationService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
 import java.time.LocalDate;
 import java.util.List;
 
 @RestController
 @RequestMapping("/notifications")
 public class NotificationController {
-    @Autowired
-    private NotificationService notificationService;
+
+    private final NotificationService notificationService;
+    private final ValidationService validationService;
+
+    public NotificationController(NotificationService notificationService, ValidationService validationService) {
+        this.notificationService = notificationService;
+        this.validationService = validationService;
+    }
 
     @PostMapping
     private ResponseEntity<NotificationResponseDto> register(@RequestBody NotificationRequestDto notificationRequest) {
+        validationService.verifyIsAdminUser();
         return ResponseEntity
                 .status(201)
                 .body(toResponse(notificationService.register(notificationRequest)));
@@ -62,6 +69,7 @@ public class NotificationController {
 
     @DeleteMapping("/{id}")
     private void deleteById(@PathVariable Long id) {
+        validationService.verifyIsAdminUser();
         notificationService.deleteById(id);
     }
